git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1439 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/hrestest.c
@@ -106,7 +106,7 @@ int main(int argc, char** argv)
       puts("RSerPool Handle Resolution Tester - Version 1.0");
       puts("===============================================\n");
       printf("Pool Handle         = %s\n",   poolHandle);
-      printf("Items               = %ums\n", items);
+      printf("Items               = %ums\n", (unsigned int)items);
       printf("Pause               = %ums\n", pause);
       printf("Max Runs            = %llu\n", maxRuns);
       printf("Statistics Interval = %ums\n", statsInterval);
